Moved config to separate file
